+++ Compiler/CodeAnalysis/ExpressionSyntax.cs
@@ -1,0 +1,6 @@
+namespace CFactorial.CodeAnalysis
+{
+    abstract class ExpressionSyntax : SyntaxNode
+    {
+    }
+}
