+++ Compiler/CodeAnalysis/NumberExpressionSyntax.cs
@@ -1,0 +1,17 @@
+namespace CFactorial.CodeAnalysis
+{
+    sealed class  NumberExpressionSyntax : ExpressionSyntax
+    {
+        public NumberExpressionSyntax(SyntaxToken numberToken)
+        {
+            NumberToken = numberToken;
+        }
+        public override SyntaxKind Kind => SyntaxKind.NumberExpression;
+        public SyntaxToken NumberToken { get; }
+
+        public override IEnumerable<SyntaxNode> GetChildren()
+        {
+            yield return NumberToken;
+        }
+    }
+}
