+++ Compiler/CodeAnalysis/SyntaxKind.cs
@@ -1,0 +1,19 @@
+namespace CFactorial.CodeAnalysis
+{
+    enum SyntaxKind
+    {
+        NumberToken,
+        WhiteSpaceToken,
+        PlusToken,
+        MinusToken,
+        StarToken,
+        SlashToken,
+        OpenParenthesisToken,
+        CloseParenthesisToken,
+        BadToken,
+        EndOfFileToken,
+        NumberExpression,
+        BinaryExpression,
+        ParenthesizedExpression
+    }
+}
