+++ Compiler/CodeAnalysis/SyntaxNode.cs
@@ -1,0 +1,8 @@
+namespace CFactorial.CodeAnalysis
+{
+    abstract class SyntaxNode
+    {
+        public abstract SyntaxKind Kind { get; }
+        public abstract IEnumerable<SyntaxNode> GetChildren();
+    }
+}
