+++ Compiler/CodeAnalysis/SyntaxToken.cs
@@ -1,0 +1,23 @@
+namespace CFactorial.CodeAnalysis
+{
+    class SyntaxToken : SyntaxNode
+    {
+        public SyntaxToken (SyntaxKind kind,int position, string text, object value)
+        {
+            Kind = kind;
+            Position = position;
+            Text = text;
+            Value = value;
+        }
+
+        public override SyntaxKind Kind { get; }
+        public int Position { get; }
+        public string Text { get; }
+        public object Value { get; }
+
+        public override IEnumerable<SyntaxNode> GetChildren()
+        {
+            return Enumerable.Empty<SyntaxNode>();
+        }
+    }
+}
